+++ src/byob.js
@@ -97,7 +97,7 @@
 
 /*global modules, CommandBlockMorph, SpriteMorph, TemplateSlotMorph, Map,
 StringMorph, Color, DialogBoxMorph, ScriptsMorph, ScrollFrameMorph, WHITE,
-Point, HandleMorph, HatBlockMorph, BlockMorph, detect, List, Process,
+Point, HandleMorph, HatBlockMorph, BlockMorph, detect, List, Process, isString,
 AlignmentMorph, ToggleMorph, InputFieldMorph, ReporterBlockMorph, StringMorph,
 nop, radians, BoxMorph, ArrowMorph, PushButtonMorph, contains, InputSlotMorph,
 ToggleButtonMorph, IDE_Morph, MenuMorph, copy, ToggleElementMorph, fontHeight,
@@ -107,7 +107,7 @@ WatcherMorph, Variable, BooleanSlotMorph, XML_Serializer, SnapTranslator*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.byob = '2020-October-07';
+modules.byob = '2020-November-27';
 
 // Declarations
 
@@ -369,6 +369,32 @@ CustomBlockDefinition.prototype.parseChoices = function (string) {
         }
     });
     return dict;
+};
+
+CustomBlockDefinition.prototype.menuSearchWords = function () {
+    // return a single string containing words that can be searched for
+    // inside my dropdown menus
+    var terms = [];
+    this.inputNames().forEach(slot => {
+        var menu = this.dropDownMenuOf(slot);
+        if (menu) {
+            if (isString(menu)) { // special menu, translates its values
+                menu = InputSlotMorph.prototype[menu](true);
+                terms.push(
+                    Object.values(menu).map(entry => {
+                        if (isNil(entry)) {return ''; }
+                        if (entry instanceof Array) {
+                            return localize(entry[0]);
+                        }
+                        return entry.toString();
+                    }).join(' ')
+                );
+            } else { // assume a dictionary, take its keys
+                terms.push(Object.keys(menu).join(' '));
+            }
+        }
+    });
+    return terms.join(' ').toLowerCase();
 };
 
 CustomBlockDefinition.prototype.isReadOnlyInput = function (inputName) {